Rename package to operations lib
